+++ lib/beacon/boot.ex
@@ -33,6 +33,7 @@ defmodule Beacon.Boot do
 
     # temporary disable module reloadin so we can populate data more efficiently
     %{mode: :manual} = Beacon.Config.update_value(site, :mode, :manual)
+    Beacon.Loader.populate_default_media(site)
     Beacon.Loader.populate_default_components(site)
     Beacon.Loader.populate_default_layouts(site)
     Beacon.Loader.populate_default_error_pages(site)