+++ lib/beacon/loader.ex
@@ -98,6 +98,10 @@ defmodule Beacon.Loader do
     site |> modules_table_name() |> :ets.match(:"$1") |> List.flatten()
   end
 
+  def populate_default_media(site) do
+    GenServer.call(worker(site), :populate_default_media, @timeout)
+  end
+
   def populate_default_components(site) do
     GenServer.call(worker(site), :populate_default_components, @timeout)
   end
@@ -258,12 +262,14 @@ defmodule Beacon.Loader do
   end
 
   def handle_info({:layout_published, %{site: site, id: id}}, config) do
+    Beacon.Content.reload_published_layout(site, id)
     reload_layout_module(site, id)
     reload_runtime_css(site)
     {:noreply, config}
   end
 
   def handle_info({:page_published, %{site: site, id: id}}, config) do
+    Beacon.Content.reload_published_page(site, id)
     reload_page_module(site, id)
     reload_runtime_css(site)
     {:noreply, config}
@@ -271,6 +277,7 @@ defmodule Beacon.Loader do
 
   def handle_info({:pages_published, site, pages}, config) do
     for %{id: id} <- pages do
+      Beacon.Content.reload_published_page(site, id)
       reload_page_module(site, id)
     end