+++ test/support/router.ex
@@ -17,6 +17,7 @@ defmodule Beacon.BeaconTest.Router do
 
   scope "/" do
     pipe_through :browser
+    beacon_site "/other", site: :not_booted
     beacon_site "/", site: :my_site
   end
 end